Deactivate app plugins when update fails prior to recovery
When an application update fails and the previous version is recovered,
the plugin infrastructure for the new app version is not fully cleaned
up during recovery. If another application update is attempted, plugin
activation/deactivation becomes out of sync and results in a KeyError
when attempting to enable the new app plugins for update.

This update will:
 - ensure that the failed app plugins are disabled prior to cleanup
   during application recovery
 - catch the KeyError if a similar situation occurs, log a message, and
   continue to cleanup the working set for the plugins

Change-Id: If58942cd9342802bfd2055152c2f2d6289054084
Closes-Bug: #1923004
Signed-off-by: Robert Church <robert.church@windriver.com>