Restart tiller on openstack pending install check
This is another attempt at fixing the same bug as the merged review
https://review.opendev.org/c/starlingx/config/+/783472 had tried, since
there were reports indicating that the bug would still occur on certain
setups.

This patch explicitly forces a tiller restart when catching the first
HelmTillerFailure exception caused by the broken pipe error, instead of
only trying to rerun the 'helm list' command, which was believed to be
a reliable workaround to the problem, but didn't solve it in every
possible scenario.

Closes-Bug: #1917308
Signed-off-by: Gustavo Santos <gustavofaganello.santos@windriver.com>
Change-Id: I38667609173ca5c6fed028f75742ae99efedf149