Support configuration of system-controller address pools post-install
During migrating a subcloud to a new central cloud, the system
controller's network configuration needs to be re-configured in the
subcloud. This commit removes the check of is_initial_config_complete
flag during modification/deletion of the system-controller-subnet and
system-controller-oam-subnet address pools. After this commit, the two
types of address pools can be re-configured after the initial
bootstrap.

Test:
Delete the system-controller-subnet and
system-controller-oam-subnet in a subcloud after bootstrap.

Change-Id: Ied68bbfd83a0cc1c3bb0fb31ee55f924353fb4b5
Story: 2008774
Task: 42291
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>