Add pull_image option to perform_app_upload
Downloading images at application upload enables subclouds to use the
central cloud registry even when the central does not have the same
application applied. This way Central would only need to upload an app
such as Openstack in order to have images available for the subclouds.

This is done by adding an extra option `-i` or `--images` that
defaults to False.

Testing:
* VirtualBox: AIO-DX (Central), AIO-SX (Subcloud)
* Upload an app to the Central Cloud using `--image` or '-i'
* option (without applying it) Upload and apply app at
* Subcloud using the DC registry

Signed-off-by: Lucas Cavalcante <lucasmedeiros.cavalcante@windriver.com>
Change-Id: I706c2bdf233617aadae8506724dde1afbbc1b35b
Closes-Bug: 1925844