Clear host config_target on upgrade migration
The host config_target is taken as a snapshot on the upgrade-start.
This can lead to config out of date condition if the N controller
issues subsequent config requests.  This is more likely in duplex
controllers, however, for consistency the config_target is reset
to track when the N+1 controller is active.

Clear the config target of all hosts on upgrade migration.  The N+1
controller will resume tracking the config by generating config_target
when configuration is issued.

Condition for upgrade activation-complete is updated to account
for potential None config_target.

Verified upgrade complete on AIO-SX and Duplex systems.

Change-Id: I4dd44e6548a45d32ab6a0b6735a04d624da7caad
Closes-Bug: 1926512
Signed-off-by: John Kung <john.kung@windriver.com>