Memory adjustment to consider memory available
In order to handle patch applications that might change the reserved
constants defined for the platform memory, there is an auto adjustment
routine that compensates the potential extra memory by decreasing the
number of hugepages configured.
However, the routine should take into account the memory available
and only decrease in case a certain threshold is reached.
For the sake of safety on memory allocation, it was considered 50%
of total memory to be enough safe guard to avoid reaching the total
memory limit as well avoiding hugepages to be removed (mainly 1G ones).
Also, vswitch hugepages allocation have been added to the calculation
of memory allocation.

Testing:
1. Patch that increases memory allocation constant
2. AIO-DX upgrade from stx 4.0 to 5.0 (in which the platform memory
reserved increased from 7000 MB to 8000MB)

Closes-Bug: 1927172

Signed-off-by: Adriano Oliveira <adriano.oliveira@windriver.com>
Change-Id: I0bb29de83268709cdae07834f537a28010d884bc