AIO-SX reboots after change OAM ip address
On HW tests, it was detected that openstack-endpoints restart was
happening at the same as the service-manager restart, this creating
a conflict that preventing SM services to reach enabled-active.
This was provoking the reboot.

The correction add a class to execute openstack-endpoint runtime
restart on the post stage on puppet, avoiding to be run as SM is
restarted.

Tested on AIO-SX, by monitoring manifest apply and validating that
no reboot happens

Closes-Bug: 1927275
Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/789946

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: I9f547fbcc73ba5fea077c764a4a9282a02ac71c6
(cherry picked from commit 3e9982bab088df428e786d5fdf2515dc21f189fa)