Check for FEC device configuration pending
SRIOV configuration at Intel ACC100 device configuration was not being
correctly applied when it followed a bunch of other configurations, like
data networks, SRIOV interfaces, etc., and soon before the unlock of the
host, because runtime manifest was not finishing before system shutdown.

"This issue came about because of the way we handle 'pci_devices' in
contrast to 'interfaces/ports'. For the interfaces/ports, when
configuring SR-IOV, the interface sriov_numvfs will represent the user
requested value, while the underlying port sriov_numvfs will represent
the actual system value. Then, before a system can be unlocked, the
values are compared for equality. This ensures that the runtime manifest
that is setting the value of sriov_numvfs has a chance to run. For
'pci_devices' like the FPGA cards we support, there is no concept of an
'upper interface'. Therefore, the value of sriov_numvfs for a pci_device
represents the system value rather than the user requested value. This
can cause issues when performing and unlock right after configuring a
pci_device with SR-IOV. There's a chance that the system populates its
hieradata and unlocks before the runtime manifest has had a chance to
configure the SR-IOV value for the pci_device." (Steven Webster)

This change appends a "/APPLYING" string to 'extra_info' field of FEC
device when it gets configured via API, and it will be removed only when
the corresponding inventory is reported back by SYSINV agent. Until
there, attempts of unlocking the host will find that sub-string at the
field and this is subject of additional semantic check.

Closes-Bug: 1927089
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: I175bc01a2a51808c4dc7b821905c7417660bf286