Limit etcd migration to swact actions
Currently the etcd migration will take place whenever
upgrade_swact_migration.py is called. Typically this is called during a
swact as part of etcd start. However the script can also be called
during etcd restart, which occurs as part of activating secured etcd.

The solution is to limit the etcd migration to controller-0. This will
ensure the action will only be taken during the swact to controller-0.

Closes-Bug: 1927508
Signed-off-by: David Sullivan <david.sullivan@windriver.com>
Change-Id: I893eeaa7ddb0b600baa051498a30ed737c688151