Specify timeout for _get_token REST API request in cert-mon
In a DC env, when management network connectivity is disrupted
for some time, cert-mon's audit greenthreads hang. This
happens because of missing timeouts in REST API requests which
cause the urlopen to remain hanging forever. As a result, the
dc-cert sync status for the subcloud  remains unknown
indefinitely and the subclouds remain out-of-sync.

This commit adds a timeout to the _get_token method to avoid
this issue.

Change-Id: Idd41cfca6b28287de8328b1ace856cc391778cac
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Closes-Bug: 1927735