Update config file update operations to be more atomic
sysinv agent iconfig_update_file() was unconditionally deleting
the file or symlink to be updated prior to updating it. This could
render the file missing if sysinv-agent is restarted.

This update copies the updated file contents to the file and
only removes symlink, if any, when needed.

A few minor logging improvements related to logs observed in LP.

Change-Id: I3c97e778b17dd0e9693a156ffe6b1c0269413f20
Closes-Bug: 1928368
Signed-off-by: John Kung <john.kung@windriver.com>