Replace applying flag by dict for FEC device check
This change replaces the latest solution to check FEC device before an
unlock action, that relied on an '/APPLYING' flag. In certain asynchro-
nous scenarios, that flag could be cleared before than expected if an
inventory report not related to the FEC device configuration came late
(that might happen when configuring a long queue of SRIOV port changes)
or by periodic sysinv report.
The solution still uses the 'extra_info' field of PCI devices, this
time "stringifying" a dictionary entry for 'expected_numvfs' that will
keep (without clearing) at that field the programmed number of VFs at
FEC device. It is then compared with the actual sriov_numvfs of device
from the inventory report, in a similar way of what is currently done
for comparing SRIOV interfaces (from database) to ports (from device).

Closes-bug: 1927089
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: I380bd66a8229a72ef1981cbefa3a0543c28d7f30