Fix skipping recovery of the old app for app updates
When app update fails and it is requested that rollback is skipped it
still reverts to the old version of the app.

Confusion here was created by the names of the building blocks for the
update logic (perform_app_rollback, perform_app_recover). In fact it is
desired to skip operations that recover the old app.

Added the missing logic for path of failed update operation.
Now both upgrades and downgrades of the app behave the same.

Tested by changing the pvc claim to trigger the armada failure.

Closes-Bug: 1928671
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I6792744257a9cb249e0b1bf99f9b78f3b27859d9