Update k8s application upgrade script for corner case
The upgrade script may report the application update incorrectly
because there is a window when querying the application version
and status during update.

In the case that the application update to the new version is
failed, the query for the app version is made before the
application recovery is triggered, and the query for the app
status is made after the application is recovered to the previous
version, the script gets incorrect information and reports the
application is updated successfully.

Update the application query in one request to eliminate this
possiblity.

Change-Id: Icbf173214de591861c21841e22359ad981453e84
Story: 2008055
Task: 42246
Signed-off-by: Angie Wang <angie.wang@windriver.com>