Fix zuul for bandit target
Some zuul nodes running bionic do not consider the
older version of pyflakes to be installable. This seems to
be a cache issue.

This fix updates the version of hacking defined in the top
level test-requirements.txt file to use a more modern version.

It also only imports yamllint if it is python 3, since the yamllint
tox target is python 3 only.

Partial-Bug: 1928978
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: Ia7aa6a296810adc0d9ba9eca701ec70f2c4be8cd