Ensure the old app plugins enabled in app recovery lifecycle
When an application update fails, the app should be recovered
to the pervious version. In the case that the new app version
is decoupled(which has the plugins infrastructure) but the old
app version is not, the old app version is not considered to be
a system knowledgeable app which causes the plugins activation
to be skipped during the subsequent app update so that an error
occurs because the lifecycle operator is still looking for the
new app version's plugins.

Update to ensure the old app plugins are enabled before armada
process during recovery and after recovery is completed. For the
apps decoupled in stx5.0 but not decoupled in stx4.0 like
nginx-ingress-controller and portieris, only reload the operators.
The particular handling for non-decoupled apps can be removed in
the release stx6.0 as all apps in stx5.0 are decoupled.

Change-Id: Ief79baac428af7f926f8721f15ded340e3cf1e44
Closes-Bug: 1929149
Signed-off-by: Angie Wang <angie.wang@windriver.com>