Calling an additional shell lint command from zuul
The '.sh' files are examined by bashate but now the linters
command will also run a shellcheck on them.

The majority of the shellcheck errors that are failing
have been suppressed, and can be updated and fixed by
subsequent submissions.

Shell scripts that do not end in .sh will also be examined
by bashate and shellcheck.

Story: 2008943
Task: 42566
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: Ie5c58ec9391ed76bfe34b3544972306341fafcc4