Fix issue in partition data migration script
The created partition dictonary partition_map is not
an ordered dict so we need to sort it by its key -
device node when iterating it to adjust the device
nodes/paths for user created extra partitions to ensure
the number of device node/path for each extra partition
is calculated correctly, otherwise the adjustments
could be messy that causes the partition DB update to
fail.

Tested AIO-SX upgrade with three additional partitions.

Change-Id: I1cb3bbfaf144a59d29633c1784b0fde80529cd71
Closes-Bug: 1892554
Signed-off-by: Angie Wang <angie.wang@windriver.com>