Fix file permissions failure during duplex upgrade abort
When issuing a downgrade for controller-0 in a duplex upgrade
abort and rollback scenario, the downgrade command was failing
because the sysinv API does not have root permissions to set
a file flag.
The fix is to use RPC so the conductor can create the flag
and allow the downgrade for controller-0 to get further.

Partial-Bug: 1929884
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: I913bcad73309fe887a12cbb016a518da93327947