Fix controller-0 downgrade failing to kill ceph
kill_ceph_storage_monitor tried to manipulate a pmon
file that does not exist in an AIO-DX environment.

We no longer invoke kill_ceph_storage_monitor in an
AIO SX or DX env.

This allows: "system host-downgrade controller-0"
to proceed in an AIO-DX environment where that second
controller (controller-0) was upgraded.

Partial-Bug: 1929884
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: I633853f75317736084feae96b5b849c601204c13