Enable N3000 FPGA reset if device is present
As part of the switch to the singular aio manifest, the reset of the
N3000 FPGA is no longer being performed at boot by the fpga agent.
Rather, it is now initiated on demand by the puppet manifest if an
N3000 device is configured for SR-IOV.  This can cause a problem on
the first configuration of the device as the PCI address of the FEC
device changes after reset.  That is, on first configuration, a user
will be modifying the device based on a PCI address that will not
exist after the device is reset.  This will cause the subsequent
enablement of SR-IOV VFs and binding to respective PF/VF drivers
by the puppet manifest to fail.

This commit fixes the issue (with the corresponding stx/stx-puppet
change) by allowing all enabled accelerated FED devices to be
recorded in hieradata so that the N3000 can be restarted if it is
detected.

The device will only be SR_IOV configured user requested number
of VFs is different than what is present on the system.
In other words, if the device has not been configured,
the requested VFs will be 0, matching what is
present on the system (and puppet will do nothing).

Closes-Bug: #1931128
Change-Id: Id30590fdabbeb3d5dfa6a618a406b41f0a31d872
Signed-off-by: Steven Webster <steven.webster@windriver.com>