Fix for shared NIC with the N3000 FPGA
An issue can occur when using an N3000 FPGA NIC with upper
vlan interfaces (oam, mgmt, etc).  Because the N3000 is reset
after the networking initialization is complete, any IPv6
addresses, or (IPv4 / IPv6) default routes configured on the
vlan interfaces will be removed.

This commit fixes the issue by creating an FPGA resource
for puppet containing the N3000 interface name and any of
the 'used by' (upper) interfaces belonging to the interface.

If these resources are present, puppet will restart the
upper interfaces after the N3000 FPGA is reset, allowing
any configured IP addresses and routes to be restored.

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/795710
Partial-Bug: #1931461
Change-Id: Ie6725f572a32bfe649962a6eadca498bcc2c9e7c
Signed-off-by: Steven Webster <steven.webster@windriver.com>