Avoid SRIOV misconfiguration at FEC device
When performing a batch SRIOV configuration (e.g. from shell script),
it may be possible that commands consecutively issued - either related
to interface(s) or a second FEC device settings - make the FEC device
configuration to be not fully applied, and then unlock fails by
semantic check.

The reason for that is hieradata might get updated by intermediate
inventory reports that could overwrite the values to be set with the
current ones, since it hasn't separate fields for both values. This
change makes 'num_vfs' to be used as the "configured" value (like it is
actually used in puppet manifest) and creates another 'sriov_vfs' that
will contain the current value for number of VFs.

BONUS: Fixed also the missing contents of 'extra_info' field when
showing the device details (output of 'system host-device-show' command)

Partial-Bug: #1931458
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: Ie87ccfe4e27fe2fbad69b3bcc8e2687817be6893