Change Docker proxy service parameters without reboot
This commit allows for changing of Docker proxy service parameters
without needing to lock/unlock hosts.

The commit enhances service parameters to determine the need to raise
the reboot required flag based on the section of a service parameter
instead of only the service.

It also fixes an issue where the reboot required flag was incorrectly
set on operations that are not reboot required. This issue means that
a previous config out-of-date alarm from a reboot required service
parameter change can be accidentally cleared by applying another
runtime manifest.

With the ability to differentiate based on service parameter section,
this commit also fixes an issue with kubernetes service parameters,
where recently kube_apiserver and kubernetes certificate service
parameters needed a reboot because they were in the same service as
kubernetes config, which needed a reboot. As a result, kube_apiserver
and kubernetes certificate service parameters accidentally became
reboot required on workers in addition to controllers. This commit
makes them not reboot required like they were supposed to be.

Change-Id: I109ed198e2d50e1f98d3be6ff8069738fdce7f07
Closes-Bug: 1931593
Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/795825
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>