N3000 FEC device config does not recover on host re-install
Problem detected on the port update as the reports, generated before
the worker node configuration, are erasing the SRIOV parameters
obtained from the previous installation. Since the reinstall is done
in a locked state, the validation between the number of VFs and the
number of PCI addresses is failing due to this update being done
from an unconfigured state and the node cannot finish the
configuration. The correction consists of modifying the validation
to consider not test the difference between interface and port sriov
parameters, if it is not the active controller, the configuration is
out-of-date, and is locked.

During the stand-by controller reinstallation the active controller
might collect inventory data prior to the FPGA reset. This reset
might relocate the device's PCI addresses creating invalid entries
on the active database due to the initial inventory report. The
correction consists in transport the N3000 reset state back to the
conductor and use this information to decide if the entry will be
incorporated on the database (true if the reset was executed).

Closes-bug: 1929301
Depends-on: https://review.opendev.org/c/starlingx/stx-puppet/+/797330

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: Ie3db6f4b13abc905ff533660196e7935239fc6fb