Kube rootca update abort - API
This change updated kube rootca update complete API to support
update-complete and update-abort. Update-abort can be triggered
by "system kube-rootca-update-abort" which provides
a way for user to exit the update before it is fully done. After
aborting the current update, a new update can be started again
as normal.

When an the update is aborted, an alarm will be raised.
The alarm serves as a reminder that a root CA update
has not been fully finished and a new update is needed to fully
update the root CA cert and all the child certs.

The alarm will be cleared when a new update starts.

This change also updated and added tox unit tests.

Story: 2008675
Task: 43066
Depends-on: https://review.opendev.org/c/starlingx/fault/+/805371
Signed-off-by: Andy Ning <andy.ning@windriver.com>
Change-Id: Ifdcdf90e7be37c1bef969a769817fbb6a140fb48