Address python3 string issues with subprocess
This patch updates our Popen and check_output calls to enable
newlines for calls that we parse or consume the output for.
Without universal_newlines=True, the output is treated as bytes
under python3 which leads to issues later where we are using it as
strings.

See https://docs.python.org/3/glossary.html#term-universal-newlines

Story: 2006796
Task: 42667

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I2086628948ca2dfb742de01923b56949ed30dfab