Update controllerconfig testenv to use python3 by default
Change pylint, flake and cover testenvs from python2.7 to python3.
Change installed version of pylint and isort for tests to be the
same as the ones used in sysinv. Previously no version was pinned
when using python3.
Add pylint to deps.

Part of this was work was neccessary to fix flake8 and pylint errors:
- Add inline pylint suppressions.
- The Request object add_data method doesn't exist anymore. Use direct
access to data field instead.
- Remove whitespace between print function and '('.

Story: 2008454
Task: 42675
Task: 42617
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: Icfae48dbbcd8efa4abb1a557a91bf1c58c5cb5bc