Fix for shared N3000 NIC with VF upper interfaces
Previous commit ec8ff34 (Fix for shared NIC with the N3000 FPGA)
enabled the restarting of any upper interfaces of an N3000 network
interface after the FPGA is reset.  This was needed to restore any
IP configuration and routes on the upper interfaces such as in the
case of a shared NIC scenario with mgmt and oam being vlan interfaces
ontop of a shared lower interface.

An issue can occur with this in the configuration scenario of the
lower N3000 device being of class PCI-SRIOV, with type 'VF' upper
interfaces. The 'VF' upper interfaces are virtual and do not have
a corresponding kernel interface.  This means an error will occur
when puppet attempts to ifdown/ifup them after the N3000 device is
reset.

This commit fixes the issue by restricting the upper interfaces in
puppet's interface fpga config resource to be only of type vlan.

Closes-Bug: #1931461
Change-Id: I280148883b33c4aef4cfd4a53dff53ca8c3f824e
Signed-off-by: Steven Webster <steven.webster@windriver.com>