py3: Fix behavior of getsitepackages()
getsitepackages() returns a list containing all global site-packages
directories. However, the method behaves differently in python3 compared
to python2. In python2, the first element of the array is
/usr/local/python2.7, however in python3 the first element of the
array is /usr/local/lib/python3.6, which does not exist when a user
is trying to upload applications.

To get around this we check to which version is running via the six
library. If python2 is running then use the old site-packages list,
otherwise use the python3 version. This can be removed after python2
support goes away. This code has been tested on both Centos8 and Debian.

Story: 2006796
Task: 42730

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I0cae10787b9c6c0f41aa3dfedb350c382cb97fbc