System host-upgrade has inadequate semantic checks
System host-upgrade prompts the user to confirm the operation on a SX
for any hostname typed, this adds validation to the inputted argument
and rejects the command if the hostname or id typed is invalid.

Also added validation to host-downgrade and kube-host-upgrade.

Tested on a SX, changing the cgts-client and the validation was
sucessfull rejecting invalid ids and hostnames.

Closes-bug: #1934806
Signed-off-by: Iago Regiani <Lago.RodriguezRegiani@windriver.com>
Change-Id: Ia7f48fee1204e4e2178ef5a954621725512cc738