System host-lldp-agent-list does not report any lldp agents
In systems with SRIOV interfaces the kernel might have more link
devices than the ones registered as host-ports. This is caused
by the SRIOV-VF interfaces configuration. The lldp daemon does list
them on its report, and the conductor should ignore those and process
only the ones with a host-port.

The correction is on the conductor's processing loop that was
returning on the first non-match, preventing the analysis of all
array elements.

Closes-Bug: 1935680

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: I436f74785b80ba544131b22dfaecc0cdfc1cabc7