py3: Fix hashlib usage
The hashlib provides some objects. Those objects have an update function
and a constructor. In Python3 both the constructor and the update
function requires an encoded bytes object instead of a str. The str is
encoded to bytes object

In Python3 the update and constructor functions of objects provided by
hashlib requires an encoded bytes.

Use the oslo_utils encodeutils.safe_encode method to encode the text as bytes.

Story: 2006796
Task: 42807

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I41684fe41a7352709d9c983039f23a1c682396e8