Use resourceVersion when re-registering expired watch
This is related to change https://bugs.launchpad.net/starlingx/+bug/1914408

When we re-register a watch on the 410/expired event from the kubernetes
watch API, it resends the last event from the watch. This causes the
cert-mon watcher to process the event as new.

From the Watch source code, the suggestion is to query the current
kubernetes 'resourceVersion' attribute for the resource under watch, and
then apply that when re-registering the watch.

Reference:
https://github.com/kubernetes-client/python-base/blob/
b4d3aad42dc23e7a6c0e5c032691f8dc385a786c/watch/watch.py#L119:

> Note that watching an API resource can expire. The method tries to
> resume automatically once from the last result, but if that last result
> is too old as well, an `ApiException` exception will be thrown with
> ``code`` 410. In that case you have to recover yourself, probably
> by listing the API resource to obtain the latest state and then
> watching from that state on by setting ``resource_version`` to
> one returned from listing.

Local testing shows that this approach avoids the replaying of last
watch events.

Related-Bug: 1914408
Story: 2008960
Task: 42791

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I7e6370807feccaad8bc6a4ebebd3252b407d68ff