Example prepararation of tox & zuul for Debian & CentOS co-existence
The plan is for both Debian and Centos to be built from
the master branch.
At runtime they have different Python environments. Different versions
of packages are required for testing.

This commit is an example of what to modify for other repos for both
zuul and tox in order to obtain test environments for both Python 2 and
Python 3.
The targeted repo for this example was stx/config/sysinv/sysinv/sysinv.
The targeted tox test environments were flake8 and pylint.

Note:
Pep8 testenv is preserved for aligning with opendev openstack zuul jobs
(e.g openstack-tox-pep8).

Story: 2008454
Task: 42842
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I70e4521eb4438a340199964fd2dbffa29ae9932b