Kube Upgrade: Raise and clear kube upgrade in progress alarm
The fm alarm 900.007 "Kubernetes upgrade in progress." is raised on start
of kubernetes upgrade and cleared on completion of kubernetes upgrade.

Prior to this commit, the 900.007 (FM_ALARM_ID_KUBE_UPGRADE_IN_PROGRESS)
alarm was not being raised. It is already in the VIM ignore list, so
it does not prevent the orchestrated kube upgrade operations.

Verified on manual and VIM orchestrated kubernetes upgrade.

Change-Id: I97a6aa034acb4725f6a7a27f58b8e8aa311ece22
Closes-Bug: 1936372
Signed-off-by: John Kung <john.kung@windriver.com>