Remove load image file in case of import error
In a scenario in which the AMQP Server lost connection, a TimeoutError
occurred on the start_import_load RPC call to sysinv conductor.
This error was not handled so the temporary file was not removed.
The current staging cleanup routine was extended to handle
TimeoutError. Also, when writing the file to disk a possible out of
space error was detected and the file removed.

Tests were performed by killing sysinv-conductor and generating the
TimeoutError on the RPC call to conductor.
Also, no space available was simulated by adding files to the
/scratch partition.
Scenario of successful load-import was also tested, as well
load-delete.

Closes-Bug: 1937168
Signed-off-by: Adriano Oliveira <adriano.oliveira@windriver.com>
Change-Id: Ic23787df9a765a9374b4dd33f9242ee000cbf75c