Add new kube_cmd_versions table and API endpoint
In the DB, a new "kube_cmd_versions" table will be created,
aside from the boilerplate stuff, the important columns
will be "kubeadm_version" and "kubelet_version".
We want to default these to the highest-available
validated version of kubernetes in the load at the time.
There will be only 1 row here and the default will be created
in the DB migration script.

In the API a new endpoint is created '/v1/kube_cmd_versions',
and with this we can get and modify the kubeadm/kubelet version.

Story: 2008972
Task: 42895
Change-Id: Ic848ea777fecab5bda7488df225c2c3e924c2fd5
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>