Improve kube-rootca-update orchestration interactions
When a kube rootca update is completed, it needs to notify DC.
DC uses those events to know when it should trigger an audit
to determine if a subcloud is out of sync for orchestration.

During the kube-rootca-update-complete stage, the health check
is invoked, so the VIM alarm needs to be added to the ignorable
alarms list.

Story: 2008675
Task: 42914
Depends-On: https://review.opendev.org/c/starlingx/fault/+/798165
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: I56b71666b9c3c372c897a589c44a6d3cd10f8317