AIO-SX PCI device swap with the same vendor/device-id, no mgmt intf
When the system is AIO-SX, if the operator needs to replace a PCI
device with the same vendor and device-id, without reinstalling,
the interface's MAC is currently preventing the acceptance of the
replacement. This results in a new entry on the interface database
to the same device.

This task implements the MAC update if the vendor and device-id
remain the same after card replacement on the same PCI address.
Not affecting the management interface as it will be handled on a
specific task within story 2009112. For non-NIC devices, no changes
were implemented in pci_device_update_by_host() since the indexing
is by PCI address, and there will be no change from sysinv point
of view.

Since there was no unit test for iport_update_by_ihost(), implemented
not only test for the new feature but also to validate the entire
method.

The test was done on Virtualbox by changing the interface MAC address
and restarting the system to check on the address update on the
database. In the lab, the MAC update was simulated by forcing the
MAC change with "ip link set address", followed by sysinv-agent
restart to generate a new report and update to the bogus MAC.
After that, the host was unlocked and the correct MAC was back on
the database. The procedure of B&R was also tested with the MAC
change after the backup was generated.

Story: 2009112
Task: 42989

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: I38ce9ace3b39b11694dd34ad1bf85e29a54d1460