Fix the timeout so as not to fail to obtain cert
The timers to wait for the https_enabled flag to change are too short
causing a timeout and not being able to generate the certificate
correctly.

The change increases the time by 30 seconds in the function
rest_api_request which is long enough so that timeout problems do not
occur. The impact on other functions is minimal. Thus, the time needed
to make the change is enough and allows changing that flag.

Possibly in the future, some alternative is needed without using fixed
values such as repeating the command with larger timeout values and
also better handling of the exception caused by the timeout.

Closes-Bug: 1940108
Signed-off-by: Alexandre Horst <alexandre.horst@windriver.com>
Change-Id: I596717ec64a6ce769ae24e9eafb6a79885ea1018