Temporary skip part of SAN certificates tests for DebianOS
In Python 3.9 versus Python3.6 on CentOS8 and Python2.7 for CentOS7
the RFC 6125 got handling improvements in standard library, which
breaks the tests.

Skip the tests for Python 3.9. Let the designers in the certificate
area implement test cases based on the capabilities and requirements
for StarlingX(I'm not aware of all the ceritificate capabilities and
requirments).

Story: 2009101
Task: 42976
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: Ifedab3efdc04299e08f4c2c807008c785615a0e0