py3: Add py39 gate for sysinv
Add python 3.9 gate for sysinv. Make the gate non-voting for now
since the unit tests do not pass for now. A seperate commit is coming
to fix the unit tests.

Update the bindep.txt to install missing packages in the ubuntu-focal
image.

Also update the requirements to fix dependencies for python3.6 and
install dependencies that are python2.7 only.

Story: 2006796
Task: 42975

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I2512e276c8517132f952cdec0affa8cea8565035