DC: ensure subcloud is online in watch event
As a subcloud is being bootstrapped, kubernetes secrets
are created before the subcloud is online.

As a result, the DCIntermediateCertRenew watch fires
well before the subcloud is available. Cert-mon attempts
to audit the subcloud as a result of the watch event,
and fails, causing the subcloud to go into a reattempt
loop. There is also some sort of issue with the keystone
token, causing 401 unauthorized errors when the subcloud
eventually does come online.

This behaviour is completely avoided if we simply
check that the subcloud is online before processing
the watch event. If the subloud is offline the event
is ignored - the subcloud will be properly audited
when cert-mon receives the subcloud online event
from dcmanager.

Story: 2008960
Task: 42969
Change-Id: I75751e24ae233c53bd00734a7d73d517868c87e8
Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>