Update the sysinv Kubernetes upgrade mechanism
Added the following kubernetes versions in sysinv:
- v1.19.13
- v1.20.9
- v1.21.3

When a k8s upgrade is first triggered we will update
the DB with the new desired version of kubeadm, write out the
hieradata, then trigger the puppet runtime manifest to update the
bind mount for kubeadm. The code that updates the bind mount is at:
https://review.opendev.org/c/starlingx/stx-puppet/+/802136

The kubelet bind mount will be updated when the
state field in the "kube_upgrade" table is
kubernetes.KUBE_UPGRADING_KUBELETS and will have the value of
the "target_version"  field in the "kube_host_upgrade" table.
In the DB, the kubelet version will be updated when the upgrade
is complete.

For clarity, only the versions that would be valid to upgrade will be
marked as "available. The other versions will show a state of unavailable.
The desired runtime version may then be chosen via sysinv as part of the
K8s upgrade procedure.  (The “system kube-upgrade-start" command already
takes a version argument; we will ensure that only one version newer is
allowed.)

Testing done on AIO-SX:
The default installed version is v1.18.1

Tried to upgrade kubernetes to the next version (v1.19.13):

- I did the first command to upgrade $ system kube-upgrade-start v1.19.13
  After this the kubeadm_version in the DB has the expected value of: 1.19.13

- The next step will be 'system kube-upgrade-download-images'
  but I had to mimic it since we don't do an actual upgrade because we don't
  have kubernetes v1.19.13 packaged yet.

- After that, I did: $ system kube-host-upgrade controller-0 control-plane
  At this step the kubeadm bind mount is updated by the puppet manifest
  https://review.opendev.org/c/starlingx/stx-puppet/+/802136
  and I checked in /etc/fstab to have the following line:
  /usr/local/kubernetes/1.19.13/stage1 /usr/local/kubernetes/current/stage1    none    rw,bind 0       0

- Then the kube-upgrade networking was skipped for the same reason as the
  kube-upgrade-download-images step before.

- When upgrading kubelet, i.e: $ system kube-host-upgrade controller-0 kubelet
  The kubelet bind mount was updated by the puppet manifest
  and /etc/fstab had got the following new line:
  /usr/local/kubernetes/1.19.13/stage2 /usr/local/kubernetes/current/stage2    none    rw,bind 0       0

- After that, I did $ system kube-upgrade-complete
  and the kubelet_version has got the value of 1.19.13 as expected in the DB

Story: 2008972
Task: 42613
Depends-On: https://review.opendev.org/c/starlingx/integ/+/804243
Change-Id: I4a580409819df6e774a7a8f340e2d94e3377ace5
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>